+++ changelog.md
@@ -13,6 +13,7 @@
 - Added loading kamehouse.js on error pages when available to replace header and footer
 - Added loading text in the header while loading the session
 - Added suspend server button to vlc player debugger
+- Added top command to admin system commands
 - Fixed error banners background positions on mobile
 - Fixed bugs